requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.3
httpx==0.27.0
python-docx==1.1.0
pdfplumber==0.11.0
pytesseract==0.3.10
pillow==10.3.0
pdf2image==1.17.0
reportlab==4.2.0
